+++ NEWS.md
@@ -1,10 +1,12 @@
 # benchlingapi 0.0.7
 
+* Added usage vignette
+* Validates credentials
 * `get_box_contents` endpoint
 
 # benchlingapi 0.0.6
 
-* drop empty entries when creating a new entity
+* Drop empty entries when creating a new entity
 
 # benchlingapi 0.0.5